- Small fix for when you save an item selection page without items selected
git-svn-id: https://plugins.svn.wordpress.org/tradetracker-store/trunk@1156956 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ menu/itemselect.php
@@ -132,8 +132,10 @@ function itemselect() {
 		$multiid = $_GET['multiid'];
 	if( isset($_POST[ $ttstoresubmit ]) && $_POST[ $ttstoresubmit ] == 'Y' ) {
 		$Tradetracker_items = $_POST['item'];
-		$query = "DELETE FROM `".$ttstoreitemtable."` WHERE `".$ttstoreitemtable."`.`storeID` = ".$multiid."";
-		$wpdb->query($query);
+		if(isset($Tradetracker_items) && $Tradetracker_items != ""){
+			$query = "DELETE FROM `".$ttstoreitemtable."` WHERE `".$ttstoreitemtable."`.`storeID` = ".$multiid."";
+			$wpdb->query($query);
+		}
 		if(isset($_POST['itemsother']) && $_POST['itemsother']!="")
 		{
 			$itemsother = explode(",",$_POST['itemsother']);
